ecommerce_api/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.9.1
django-redis==5.0.0
channels==3.0.4
channels-redis==3.3.0
django-filter==2.4.0